+++ unsafe_extensions.go
@@ -11,6 +11,7 @@ import (
 )
 
 var UnsafePayloadKind pbbstream.Protocol = pbbstream.Protocol_UNKNOWN
+var UnsafeJsonBytesEncoder = "hex"
 
 // UnsafeResolveReaderNodeStartBlock is a function that resolved the reader node start block num, by default it simply
 // returns the value of the 'reader-node-start-block-num'. However, the function may be overwritten in certain chains